requirements.txt
langchain>=0.0.300
openai>=0.27.0
pinecone-client>=1.2.0
faiss-cpu
PyPDF2>=3.0.0
pydantic>=1.10
python-dotenv
PyYAML>=6.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
aiofiles
jinja2
numpy>=1.26
requests>=2.31
